refactor(app-graphql-playground): move AuthenticatedPlaygroundClient to playgroundClient feature folder
