+++ packages/app-graphql-playground/src/features/playgroundClient/index.ts
@@ -1,3 +1,4 @@
 export { PlaygroundClient } from "./abstractions.js";
 export type { IPlaygroundClient } from "./abstractions.js";
 export { PlaygroundClientImpl } from "./PlaygroundClient.js";
+export { AuthenticatedPlaygroundClient } from "./AuthenticatedPlaygroundClient.js";
