+++ packages/app-graphql-playground/src/features/tabRegistry/index.ts
@@ -1,4 +1,3 @@
 export { PlaygroundTabRegistry } from "./abstractions.js";
 export type { IPlaygroundTabRegistry } from "./abstractions.js";
 export type { IPlaygroundTabDefinition } from "./abstractions.js";
-export { AuthenticatedPlaygroundClient } from "./AuthenticatedPlaygroundClient.js";
