+++ packages/app-headless-cms/src/admin/features/playgroundTabs/CmsPlaygroundTabs.ts
@@ -3,7 +3,7 @@ import { AuthenticationContext } from "@webiny/app-admin/features/security/Authe
 import { TenantContext } from "@webiny/app-admin/features/tenancy/abstractions.js";
 import { PlaygroundTabRegistry } from "@webiny/app-graphql-playground/features/tabRegistry/index.js";
 import { PlaygroundClientImpl } from "@webiny/app-graphql-playground/features/playgroundClient/index.js";
-import { AuthenticatedPlaygroundClient } from "@webiny/app-graphql-playground/features/tabRegistry/index.js";
+import { AuthenticatedPlaygroundClient } from "@webiny/app-graphql-playground/features/playgroundClient/index.js";
 import type { PlaygroundClient } from "@webiny/app-graphql-playground/features/playgroundClient/index.js";
 import { manageQuery } from "./queries/placeholder.manage.graphql.js";
 import { readQuery } from "./queries/placeholder.read.graphql.js";
